Updated some functionality need to focus on handling cash

+++ core_calc.cpp
@@ -2,8 +2,8 @@
 #include "core_calc.h"
 
 // Taking our bond input to calculate the periodic coupon payment.
-int pcp(const Bond *bond) {
-    long int c_pay = bond->c_rate * bond->pval / 100000;
+long long pcp(const Bond *bond) {
+    long long c_pay = bond->c_rate * bond->pval / 100000;
     std::cout << c_pay << std::endl;
     return 0;
 }
